Move protect and restrict middlewares to controllers
